requirements.txt
click
colorama
pyperclip
tabulate
pycryptodome